+++ src/components/Common/Alarm/types.ts
@@ -1,0 +1,11 @@
+export interface Alarm extends Author {
+  _id?: string;
+  post?: string;
+  author?: Author;
+  comment?: string;
+}
+
+interface Author {
+  fullName?: string;
+  image?: string;
+}
